+++ src/main/kotlin/com/github/blarc/sops/intellij/plugin/settings/ProjectService.kt
@@ -3,16 +3,12 @@ package com.github.blarc.sops.intellij.plugin.settings
 import com.github.blarc.sops.intellij.plugin.SopsWrapper
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.components.Service
+import com.intellij.openapi.project.Project
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
 
-@Service(Service.Level.APP)
-class AppService(private val cs: CoroutineScope) {
-
-    companion object {
-        val instance: AppService
-            get() = ApplicationManager.getApplication().getService(AppService::class.java)
-    }
+@Service(Service.Level.PROJECT)
+class ProjectService(private val project: Project, private val cs: CoroutineScope) {
 
     fun version(
         sopsPath: String,
@@ -20,7 +16,7 @@ class AppService(private val cs: CoroutineScope) {
         onError: suspend (message: String) -> Unit = {}
     ) {
         cs.launch {
-            SopsWrapper.version(sopsPath, onSuccess, onError)
+            SopsWrapper.version(sopsPath, project, onSuccess, onError)
         }
     }
 }